anki_generator/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.35.6
outcome==1.3.0.post0
packaging==24.1
pydantic==2.7.4
pydantic_core==2.18.4
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.22.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.2
webdriver-manager==4.0.1
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
